#SPD-1628: adding changes to execute test run successfully with threads
